+++ refacFabela/src/main/java/com/refacFabela/service/CajaService.java
@@ -1,5 +1,6 @@
 package com.refacFabela.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.refacFabela.dto.BalanceCajaDto;
@@ -11,7 +12,7 @@ public interface CajaService {
 	public TwCaja obtenerCajaActiva();
 	public List<TwCaja> obteneroCajas();
 	public BalanceCajaDto obtenerBalanceCaja(Long nIdCaja);
-	public TwCaja nuevaCaja(Double saldoInicial, Long nIdUsuario);
+	public TwCaja nuevaCaja(BigDecimal saldoInicial, Long nIdUsuario);
 	public TwCaja consultaCaja( Long nidCaja);
 	public List<TwGasto> obteberGastosCaja( Long nIdCaja);
 	public TwGasto guardarGasto( TwGasto twGasto);
